#2633 Split PatchDelegate into 2 classes, 1 without jdt.core deps.
